requirements.txt
Flask==3.0.2
flask-cors==4.0.0
psycopg2-binary==2.9.9
httpx==0.27.0
uuid6==2024.1.12
python-dotenv==1.0.1
gunicorn==21.2.0
gevent==24.2.1